refactor(all): capitalize folder names.
