+++ src/features/City/WalkingMayor.tsx
