+++ src/features/City/housesSvg/HouseCityHall.tsx
