+++ src/features/City/housesSvg/HouseGround.tsx
