+++ src/features/City/housesSvg/HouseOfClothes.tsx
