+++ src/features/City/housesSvg/HouseOfDrinks.tsx
