+++ src/features/City/housesSvg/HouseOfFood.tsx
