+++ src/features/City/housesSvg/HouseOfPlants.tsx
