+++ src/features/Statistics/BarChart.tsx
