+++ src/features/Statistics/GeneralAnalysis.tsx
